+++ pkg/local_object_storage/shard/bench_test.go
@@ -1,4 +1,4 @@
-package blobstor_test
+package shard_test
 
 import (
 	"crypto/rand"