+++ pkg/local_object_storage/shard/lock_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	objectcore "github.com/nspcc-dev/neofs-node/pkg/core/object"
-	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/fstree"
 	meta "github.com/nspcc-dev/neofs-node/pkg/local_object_storage/metabase"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard"
@@ -24,11 +23,8 @@ func TestShard_Lock(t *testing.T) {
 	rootPath := t.TempDir()
 	opts := []shard.Option{
 		shard.WithLogger(zap.NewNop()),
-		shard.WithBlobStorOptions(
-			blobstor.WithStorages(blobstor.SubStorage{
-				Storage: fstree.New(
-					fstree.WithPath(filepath.Join(rootPath, "blob"))),
-			}),
+		shard.WithBlobstor(fstree.New(
+			fstree.WithPath(filepath.Join(rootPath, "fstree"))),
 		),
 		shard.WithMetaBaseOptions(
 			meta.WithPath(filepath.Join(rootPath, "meta")),