+++ pkg/local_object_storage/shard/status.go
@@ -1,7 +1,6 @@
 package shard
 
 import (
-	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor"
 	meta "github.com/nspcc-dev/neofs-node/pkg/local_object_storage/metabase"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/writecache"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
@@ -12,10 +11,19 @@ import (
 // Metabase, and Writecache. Additionally, it includes a slice of errors that may have
 // occurred at the object level.
 type ObjectStatus struct {
-	Blob       blobstor.ObjectStatus
+	Blob       StorageObjectStatus
 	Metabase   meta.ObjectStatus
 	Writecache writecache.ObjectStatus
 	Errors     []error
+}
+
+// StorageObjectStatus represents the status of the object in the storage,
+// containing the type and path of the storage and an error if it
+// occurred.
+type StorageObjectStatus struct {
+	Type  string
+	Path  string
+	Error error
 }
 
 // ObjectStatus returns the status of the object in the Shard. It contains status
@@ -23,8 +31,13 @@ type ObjectStatus struct {
 func (s *Shard) ObjectStatus(address oid.Address) (ObjectStatus, error) {
 	var res ObjectStatus
 	var err error
-	res.Blob, err = s.blobStor.ObjectStatus(address)
-	if res.Blob.Type != "" {
+
+	_, err = s.blobStor.Get(address)
+	if err == nil {
+		res.Blob = StorageObjectStatus{
+			Type: s.blobStor.Type(),
+			Path: s.blobStor.Path(),
+		}
 		res.Errors = append(res.Errors, err)
 		res.Metabase, err = s.metaBase.ObjectStatus(address)
 		res.Errors = append(res.Errors, err)